refactor(RenderContainer): remove fallback for Portals
ReactDOM.createPortal exists since react-dom@16

BREAKING CHANGE: drop support for react-dom@15
