+++ packages/retail-ui/components/Popup/__tests__/Popup-test.tsx
@@ -6,14 +6,9 @@ import RenderContainer from '../../RenderContainer/RenderContainer';
 import ZIndex from '../../ZIndex';
 import { Transition } from 'react-transition-group';
 import LifeCycleProxy from '../../internal/LifeCycleProxy';
-import ReactDOM from 'react-dom';
-import { RenderInnerContainer as RenderContainerNative } from '../../RenderContainer/RenderContainerNative';
-import { RenderInnerContainer as RenderContainerFallback } from '../../RenderContainer/RenderContainerFallback';
+import { RenderInnerContainer } from '../../RenderContainer/RenderInnerContainer';
 import { Nullable } from '../../../typings/utility-types';
 import { ReactComponentLike } from 'prop-types';
-
-const HAS_BUILTIN_PORTAL = !!ReactDOM.createPortal;
-const RenderInnerContainer = HAS_BUILTIN_PORTAL ? RenderContainerNative : RenderContainerFallback;
 
 const openPopup = async (wrapper: ReactWrapper<PopupProps, PopupState, Popup>) =>
   new Promise(async resolve => {
