+++ packages/retail-ui/components/RenderContainer/RenderContainer.tsx
@@ -1,13 +1,8 @@
 import * as React from 'react';
-import ReactDOM from 'react-dom';
 import { Nullable } from '../../typings/utility-types';
-import { RenderInnerContainer as RenderContainerFallback } from './RenderContainerFallback';
-import { RenderInnerContainer as RenderContainerNative } from './RenderContainerNative';
+import { RenderInnerContainer } from './RenderInnerContainer';
 import { RenderContainerProps } from './RenderContainerTypes';
 import { getRandomID } from '../../lib/utils';
-
-const HAS_BUILTIN_PORTAL = !!ReactDOM.createPortal;
-const RenderInnerContainer = HAS_BUILTIN_PORTAL ? RenderContainerNative : RenderContainerFallback;
 
 export class RenderContainer extends React.Component<RenderContainerProps> {
   private static getRootId = () => getRandomID();
