+++ packages/retail-ui/components/RenderContainer/RenderInnerContainer.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Nullable } from '../../typings/utility-types';
 import { PortalProps, RenderContainerProps } from './RenderContainerTypes';
 
-interface RenderContainerNativeProps extends RenderContainerProps {
+interface RenderInnerContainerProps extends RenderContainerProps {
   domContainer: Nullable<HTMLElement>;
   rootId: string;
 }
@@ -12,7 +12,7 @@ export function Portal(props: PortalProps) {
   return <noscript data-render-container-id={props.rt_rootID} />;
 }
 
-export class RenderInnerContainer extends React.Component<RenderContainerNativeProps> {
+export class RenderInnerContainer extends React.Component<RenderInnerContainerProps> {
   public render() {
     if (this.props.children) {
       if (!this.props.domContainer) {
